Fixed bugs

+++ hypr/.config/hypr/scripts/toggle-waybar.sh
@@ -1,8 +1,14 @@
 #!/usr/bin/env bash
 
-if pgrep -f "^waybar$" >/dev/null; then
-  pkill -USR1 -f "^waybar$"
-else
-  waybar &
-fi
+GAMEMODE_STATUS=$(cat ~/.config/hypr/vars/gamemode)
 
+if [ "$GAMEMODE_STATUS" == "on" ]; then
+    notify-send -e "Please Exit Gamemode"
+    exit
+else
+    if pgrep -f "^waybar$" >/dev/null; then
+        pkill -USR1 -f "^waybar$"
+    else
+        waybar &
+    fi
+fi
